requirements.txt
aiohttp>=3.13
mss>=10.1
Pillow>=12.0
pyautogui>=0.9.54
opencv-python>=4.8
easyocr>=1.7.2
numpy>=2.0

# Optional (audio transcription)
faster-whisper>=1.2.1
